+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* ✅ test(slices-fill): fill test cases and benchmarks added
 * 🎉 feat(slices-fill): fill function added
 * 🐛 fix: bug in drop_by_test fixed
 * ✅ test(slices-drop_by): drop_by test cases and benchmarks added